clean up the files
