+++ Makefile
@@ -7,13 +7,12 @@ SOURCE_FILES := \
 					reader/reader.c \
 					builtins/builtin_echo.c \
 					builtins/builtin_pwd.c \
-					lexer/lexer.c \
 					lexer/star_expand_file_utils.c \
 					lexer/star_expand_match_utils.c \
 					lexer/star_expand.c \
 					lexer/token_composer.c \
 					lexer/token_string_utils.c \
-					lexer/tokenize.c \
+					lexer/tokenizer.c \
 					builtins/builtin_env.c \
 					builtins/builtin_unset.c \
 					builtins/builtin_export.c \
